workers_service/src/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic-settings==2.0.3
cassandra-driver==3.29.1
python-multipart


